+++ nakadi-types/src/helpers.rs
@@ -4,6 +4,9 @@ pub const NAKADION_PREFIX: &str = "NAKADION";
 
 macro_rules! env_funs {
     ($var:expr) => {
+        #[doc="The default name of the environment variable for this type.\n"]
+        pub const ENV_TYPE_NAME: &'static str = &$var;
+
         #[doc="Initialize from the environment.\n"]
         #[doc="Returns `None` if the value was not found and fails if the value could not be parsed.\n"]
         #[doc="The name of the environment variable is \"NAKADION_"]
@@ -50,6 +53,15 @@ macro_rules! env_funs {
         }
 
         #[doc="Initialize from the environment.\n"]
+        #[doc="Returns `None` if the value was not found and fails if the value could not be parsed.\n"]
+        #[doc="The name of the environment variable is \""]
+        #[doc=$var]
+        #[doc="\""]
+        pub fn try_from_env_type_named() -> Result<Option<Self>, $crate::Error> {
+            Self::try_from_env_named(Self::ENV_TYPE_NAME)
+        }
+
+        #[doc="Initialize from the environment.\n"]
         #[doc="Fails if the value was not found or if the value could not be parsed.\n"]
         #[doc="The name of the environment variable is \"NAKADION_"]
         #[doc=$var]
@@ -85,6 +97,16 @@ macro_rules! env_funs {
         }
 
         #[doc="Initialize from the environment.\n"]
+        #[doc="Fails if the value was not found or if the value could not be parsed.\n"]
+        #[doc="The name of the environment variable is \""]
+        #[doc=$var]
+        #[doc="\""]
+        pub fn from_env_type_named() -> Result<Self, $crate::Error> {
+            Self::from_env_named(Self::ENV_TYPE_NAME)
+        }
+
+
+        #[doc="Initialize from the environment.\n"]
         #[doc="Returns `None` if the value could not be read for any reason.\n"]
         #[doc="The name of the environment variable is \"NAKADION_"]
         #[doc=$var]
@@ -110,6 +132,15 @@ macro_rules! env_funs {
         #[doc="The name of the environment variable is `var_name`."]
         pub fn from_env_opt_named<T: AsRef<str>>(var_name: T) -> Option<Self> {
             Self::from_env_named(var_name.as_ref()).ok()
+        }
+
+        #[doc="Initialize from the environment.\n"]
+        #[doc="Returns `None` if the value could not be read for any reason.\n"]
+        #[doc="The name of the environment variable is \"NAKADION_"]
+        #[doc=$var]
+        #[doc="\""]
+        pub fn from_env_opt_type_named() -> Option<Self> {
+            Self::from_env_opt_named(Self::ENV_TYPE_NAME)
         }
     };
 }