+++ nakadion/CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.21.0] - 2020-04-28
+
+### CHANGED
+
+- renamed `SubscriptionCursorWithoutToken` to `EventTypeCursor` because the name was too long
+
+### ADDED
+
+- Default env name for types as constants
+- functions to retrieve values from environment by default type name
+
 ## [0.20.4] - 2020-04-27
 
 ### CHANGED